Update dependencies & clean up codebase
