+++ Cargo.toml
@@ -1,0 +1,13 @@
+[workspace]
+members = [
+  "src-tauri",
+  "src-tauri/bmm-lib",
+]
+resolver = "2"
+
+# Centralize shared deps and the local path for bmm-lib
+[workspace.dependencies]
+bmm-lib = { path = "src-tauri/bmm-lib", version = "0.2.7" }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+
