+++ src-tauri/Cargo.toml
@@ -24,7 +24,7 @@ tauri-plugin-shell = "2.3.1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 tauri-plugin-dialog = "2.3.3"
-bmm-lib = { path = "bmm-lib" }
+bmm-lib = { workspace = true }
 dirs = "6.0.0"
 tauri-plugin-fs = "2"
 log = "0.4.27"
