+++ src-tauri/bmm-lib/Cargo.toml
@@ -21,7 +21,7 @@ libflate = "2.1.0"
 log = "0.4.27"
 regex = "1.11.2"
 reqwest = { version = "0.12.23", features = ["json"] }
-rusqlite = { version = "0.36.0", features = ["bundled"] }
+rusqlite = { version = "0.37.0", features = ["bundled"] }
 serde = { version = "1.0.219", features = ["serde_derive"] }
 serde_json = "1.0.143"
 serde_repr = "0.1.20"
@@ -36,11 +36,11 @@ tempfile = "3.21.0"
 scopeguard = "1.2.0"
 mockito = "1.7.0"
 tokio = { version = "1.47.1", features = ["full", "test-util"] }
-rstest = "0.25.0"
+rstest = "0.26.1"
 
 [target.'cfg(windows)'.dependencies]
 winreg = "0.55.0"
-sysinfo = "0.35.2"
+sysinfo = "0.37.0"
 
 [target.'cfg(unix)'.dependencies]
 libproc = "0.14.10"
